# Fix quantization and dithering.