… src/ImageSharp/Primitives/DenseMatrix{T}.cs
… src/ImageSharp/Processing/Extensions/Binarization/BinaryDiffuseExtensions.cs
… src/ImageSharp/Processing/Extensions/Binarization/BinaryDitherExtensions.cs
… src/ImageSharp/Processing/Extensions/Dithering/DiffuseExtensions.cs
… src/ImageSharp/Processing/Extensions/Dithering/DitherExtensions.cs
… src/ImageSharp/Processing/Extensions/Quantization/QuantizeExtensions.cs
… src/ImageSharp/Processing/KnownDiffusers.cs
… src/ImageSharp/Processing/KnownDitherers.cs
… src/ImageSharp/Processing/KnownDitherings.cs
… src/ImageSharp/Processing/Processors/Binarization/BinaryErrorDiffusionProcessor.cs
… src/ImageSharp/Processing/Processors/Binarization/BinaryErrorDiffusionProcessor{TPixel}.cs
… src/ImageSharp/Processing/Processors/Binarization/BinaryOrderedDitherProcessor.cs
… src/ImageSharp/Processing/Processors/Binarization/BinaryOrderedDitherProcessor{TPixel}.cs
… src/ImageSharp/Processing/Processors/Dithering/AtkinsonDither.cs
… src/ImageSharp/Processing/Processors/Dithering/BayerDither2x2.cs
… src/ImageSharp/Processing/Processors/Dithering/BayerDither4x4.cs
… src/ImageSharp/Processing/Processors/Dithering/BayerDither8x8.cs
… src/ImageSharp/Processing/Processors/Dithering/BurksDither.cs
… src/ImageSharp/Processing/Processors/Dithering/DitherTransformColorBehavior.cs
… src/ImageSharp/Processing/Processors/Dithering/DitherType.cs
… src/ImageSharp/Processing/Processors/Dithering/ErroDither.KnownTypes.cs
… src/ImageSharp/Processing/Processors/Dithering/ErrorDiffuser.cs
… src/ImageSharp/Processing/Processors/Dithering/ErrorDiffusionPaletteProcessor.cs
… src/ImageSharp/Processing/Processors/Dithering/ErrorDiffusionPaletteProcessor{TPixel}.cs
… src/ImageSharp/Processing/Processors/Dithering/ErrorDither.cs
… src/ImageSharp/Processing/Processors/Dithering/EuclideanPixelMap{TPixel}.cs
… src/ImageSharp/Processing/Processors/Dithering/FloydSteinbergDither.cs
… src/ImageSharp/Processing/Processors/Dithering/IDither.cs
… src/ImageSharp/Processing/Processors/Dithering/IErrorDiffuser.cs
… src/ImageSharp/Processing/Processors/Dithering/IOrderedDither.cs