### Prerequisites

- [x] I have written a descriptive pull-request title
- [x] I have verified that there are no overlapping [pull-requests](https://github.com/SixLabors/ImageSharp/pulls) open
- [x] I have verified that I am following matches the existing coding patterns and practice as demonstrated in the repository. These follow strict Stylecop rules :cop:.
- [x] I have provided test coverage for my change (where applicable)

### Description
A deep investigation into our quantization and dithering algorithms yielded issues we needed to fix.

- Error Diffusion algorithm offset was incorrect.
- Ordered Dithering algorithm was completely incorrect. 
- Quantization could not utilize both algorithms.
- Quantization could not be isolated to a given bounds.
- `Color.Transparent` did not match W3C specification, rather copied incorrect System.Drawing value.
- Tests were leaking memory.

This PR:

- Updates Error Diffusion algorithm with correct implementation.
- Updates Ordered Dithering algorithm with correct implementation.
- Merges `IErrorDiffusion` and `IOrderedDither` APIs into new `IDither` api.
- Updates and simplifies dithering processor to use new API
- Updates and simplifies format quantization and quantization processor to use new API.
- Introduces new `QuantizerOptions` for easier bundling of options.
- Introduces new dither scaling options to allow variable dithering.
- Fixed `Color.Transparent` value.
- Improves performance across the board.
- Fixes memory leak in Bitmap Encoder
- Fixes memory leak in Quantizer tests.

### Examples
Here's an example of the new dithering output vs the old using `BayerDither8x8` ordered dithering. Note the much improved handling of gradients.

**Old**
![DitherFilter_WorksWithAllDitherers_Bike_BayerDither8x8-Old](https://user-images.githubusercontent.com/385879/74604186-8a027e80-510f-11ea-86c2-bd5704365fa4.png)

**New**
![DitherFilter_WorksWithAllDitherers_Bike_BayerDither8x8-New](https://user-images.githubusercontent.com/385879/74604208-b4543c00-510f-11ea-9e5a-37c6584cc5a5.png)

With the new `DitherScale` options built into `QuantizerOptions` it's possible to incrementally dither output.

Here's incremental quantization and dithering of a sample image. I've used the websafe palette to accentuate differences.

**Original**
![david](https://user-images.githubusercontent.com/385879/74604246-1614a600-5110-11ea-8fc8-e9a67c3d6318.png)

**0%**
![FloydSteinbergDither 0%](https://user-images.githubusercontent.com/385879/74604326-0ba6dc00-5111-11ea-9298-7b29302fdfc5.png)

**25%**
![FloydSteinbergDither 25%](https://user-images.githubusercontent.com/385879/74604324-0a75af00-5111-11ea-9c51-3931310183ce.png)

**50%**
![FloydSteinbergDither 50%](https://user-images.githubusercontent.com/385879/74604323-09448200-5111-11ea-81bf-6203db14f900.png)

**75%**
![FloydSteinbergDither 75%](https://user-images.githubusercontent.com/385879/74604325-0b0e4580-5111-11ea-86fd-16892de03cef.png)

**100%**
![FloydSteinbergDither 100%](https://user-images.githubusercontent.com/385879/74604327-0c3f7280-5111-11ea-9385-c841fb843d91.png)

### Benchmarks

_Note: Previous benchmarks did not have Core 3.1 versions. Memory increase is due to changing of color lookup caching strategy._

``` bash
BenchmarkDotNet=v0.12.0, OS=Windows 10.0.18363
Intel Core i7-8650U CPU 1.90GHz(Kaby Lake R), 1 CPU, 8 logical and 4 physical cores
.NET Core SDK = 3.1.101

[Host]     : .NET Core 3.1.1 (CoreCLR 4.700.19.60701, CoreFX 4.700.19.60801), X64 RyuJIT
Job-OJKYBT : .NET Framework 4.8 (4.8.4121.0), X64 RyuJIT
Job-RZWLFP : .NET Core 2.1.15 (CoreCLR 4.6.28325.01, CoreFX 4.6.28327.02), X64 RyuJIT
Job-NUYUQV : .NET Core 3.1.1 (CoreCLR 4.700.19.60701, CoreFX 4.700.19.60801), X64 RyuJIT

IterationCount=3  LaunchCount=1  WarmupCount=3
```
**Before**
|    Method |  Job | Runtime |      Mean |     Error |    StdDev | Gen 0 | Gen 1 | Gen 2 | Allocated |
|---------- |----- |-------- |----------:|----------:|----------:|------:|------:|------:|----------:|
| DoDiffuse |  Clr |     Clr | 124.93 ms | 33.297 ms | 1.8251 ms |     - |     - |     - |      2 KB |
| DoDiffuse | Core |    Core |  89.63 ms |  9.895 ms | 0.5424 ms |     - |     - |     - |   1.91 KB |

**After**
|    Method |       Runtime |      Mean |     Error |    StdDev | Gen 0 | Gen 1 | Gen 2 | Allocated |
|---------- |-------------- |----------:|----------:|----------:|------:|------:|------:|----------:|
| DoDiffuse |    .NET 4.7.2 | 30.535 ms | 19.217 ms | 1.0534 ms |     - |     - |     - |  26.25 KB |
|  DoDither |    .NET 4.7.2 | 14.174 ms |  1.625 ms | 0.0891 ms |     - |     - |     - |  31.38 KB |
| DoDiffuse | .NET Core 2.1 | 15.984 ms |  3.686 ms | 0.2020 ms |     - |     - |     - |  25.98 KB |
|  DoDither | .NET Core 2.1 |  8.646 ms |  1.635 ms | 0.0896 ms |     - |     - |     - |  28.99 KB |
| DoDiffuse | .NET Core 3.1 | 16.235 ms |  9.612 ms | 0.5269 ms |     - |     - |     - |  25.96 KB |
|  DoDither | .NET Core 3.1 |  8.429 ms |  1.270 ms | 0.0696 ms |     - |     - |     - |  31.61 KB |







<!-- Thanks for contributing to ImageSharp! -->